+++ HCMS-Api/Components/DMS/ESS/SubDepartmentComponent.cs
@@ -164,7 +164,7 @@ public class SubDepartmentComponent
             return new SubDepartmentReadDto
             {
                 Id = row.Field<int>("Id"),
-                CompanyId = row.Field<Int64>("CompanyId"),
+                CompanyId = row.Field<int>("CompanyId"),
                 Company = row.Field<string>("Company"), 
                 Code = row.Field<string>("Code"),
                 Name = row.Field<string>("Name"),
@@ -283,7 +283,7 @@ public class SubDepartmentComponent
                 .Select(row => new SubDepartmentReadDto
                 {
                     Id = row.Table.Columns.Contains("Id") ? row.Field<int>("Id") : 0,
-                    CompanyId = row.Field<Int64>("CompanyId"),
+                    CompanyId = row.Field<int>("CompanyId"),
                     Company = row.Field<string>("Company"),
                     Code = row.Table.Columns.Contains("Code") ? row.Field<string>("Code") : string.Empty,
                     Name = row.Table.Columns.Contains("Name") ? row.Field<string>("Name") : string.Empty,
@@ -374,7 +374,7 @@ public class SubDepartmentComponent
             return new SubDepartmentReadDto
             {
                 Id = row.Table.Columns.Contains("Id") ? row.Field<int>("Id") : 0,
-                CompanyId = row.Field<Int64>("CompanyId"),
+                CompanyId = row.Field<int>("CompanyId"),
                 Company = row.Field<string>("Company"),
                 Code = row.Table.Columns.Contains("Code") ? row.Field<string>("Code") : string.Empty,
                 Name = row.Table.Columns.Contains("Name") ? row.Field<string>("Name") : string.Empty,
@@ -424,7 +424,7 @@ public class SubDepartmentComponent
                 .Select(row => new SubDepartmentReadDto
                 {
                     Id = row.Table.Columns.Contains("Id") ? row.Field<int>("Id") : 0,
-                    CompanyId = row.Field<Int64>("CompanyId"),
+                    CompanyId = row.Field<int>("CompanyId"),
                     Company = row.Field<string>("Company"),
                     Code = row.Table.Columns.Contains("Code") ? row.Field<string>("Code") : string.Empty,
                     Name = row.Table.Columns.Contains("Name") ? row.Field<string>("Name") : string.Empty,
@@ -544,7 +544,7 @@ public class SubDepartmentComponent
             return new SubDepartmentReadDto
             {
                 Id = row.Field<int>("Id"),
-                CompanyId = row.Field<Int64>("CompanyId"),
+                CompanyId = row.Field<int>("CompanyId"),
                 Company = row.Field<string>("Company"),
                 Code = row.Field<string>("Code"), // 🔒 immutable
                 Name = row.Field<string>("Name"),
